+++ connecteur/purge/Purge.class.php
@@ -73,10 +73,11 @@ class Purge extends Connecteur {
 
     }
 
-	/**
-	 * @return bool
-	 * @throws Exception
-	 */
+    /**
+     * @return bool
+     * @throws UnrecoverableException
+     * @throws Exception
+     */
 	public function purger(){
 		if (! $this->isActif()){
 			throw new UnrecoverableException("Le connecteur n'est pas actif");
@@ -111,7 +112,8 @@ class Purge extends Connecteur {
 				$document_info['id_e'],
 				$document_info['id_d'],
 				0,
-				$etat_cible
+				$etat_cible,
+                $this->connecteurConfig->get('verrou')
 			);
 			$this->lastMessage .= get_hecho("{$document_info['id_d']} - {$document_info['titre']} - {$document_info['last_action_date']}") . "<br/>";
 		}